pyproject.toml
[project]
name = "price-monitor"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "selectolax>=0.3",
    "apscheduler>=3.10",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "jinja2>=3.1",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "plyer>=2.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.2", "pytest-asyncio>=0.23"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-p no:launch_ros -p no:launch_testing -p no:launch_pytest -p no:ament_lint -p no:ament_copyright -p no:ament_flake8 -p no:ament_pep257 -p no:ament_xmllint"
